Style: Restructured directory structures & formatted entities
