+++ src/main/java/com/dope/breaking/domain/hashtag/CommentHashtag.java
@@ -1,21 +1,23 @@
-package com.dope.breaking.domain.Hashtag;
-
-
-import com.dope.breaking.domain.Post;
+package com.dope.breaking.domain.hashtag;
+import com.dope.breaking.domain.comment.Comment;
+import lombok.Getter;
 
 import javax.persistence.*;
 
 @Entity
-public class PostHashtag {
+@Getter
+public class CommentHashtag {
+
     @Id @GeneratedValue
     @Column(name =  "POST_HASHTAG_ID")
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn (name = "POST_ID")
-    private Post post;
+    @JoinColumn (name = "COMMENT_ID")
+    private Comment comment;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn (name = "HASHTAG_ID")
     private Hashtag hashtag;
+
 }
