+++ src/main/java/com/dope/breaking/domain/hashtag/Hashtag.java
@@ -1,4 +1,6 @@
-package com.dope.breaking.domain.Hashtag;
+package com.dope.breaking.domain.hashtag;
+
+import lombok.Getter;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -6,6 +8,7 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 
 @Entity
+@Getter
 public class Hashtag {
 
     @Id @GeneratedValue
