+++ src/main/java/com/dope/breaking/domain/media/MediaType.java
@@ -1,4 +1,4 @@
-package com.dope.breaking.domain;
+package com.dope.breaking.domain.media;
 
 public enum MediaType {
     PHOTO, VIDEO;
