+++ src/main/java/com/dope/breaking/domain/post/Location.java
@@ -1,10 +1,14 @@
-package com.dope.breaking.domain;
+package com.dope.breaking.domain.post;
 
 import javax.persistence.Embeddable;
 
 @Embeddable
 public class Location {
+
     private String region;
+
     private Double longitude;
+
     private Double latitude;
+
 }
