+++ src/main/java/com/dope/breaking/domain/post/PostType.java
@@ -1,4 +1,4 @@
-package com.dope.breaking.domain;
+package com.dope.breaking.domain.post;
 
 public enum PostType {
     EXCLUSIVE, CHARGED, FREE;
