+++ src/main/java/com/dope/breaking/domain/user/Bookmark.java
@@ -1,9 +1,15 @@
-package com.dope.breaking.domain;
+package com.dope.breaking.domain.user;
+
+import com.dope.breaking.domain.post.Post;
+import com.dope.breaking.domain.user.User;
+import lombok.Getter;
 
 import javax.persistence.*;
 
 @Entity
+@Getter
 public class Bookmark {
+
     @Id @GeneratedValue
     @Column(name = "BOOKMARK_ID")
     private Long id;
@@ -15,4 +21,5 @@ public class Bookmark {
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn (name = "POST_ID")
     private Post post;
+
 }
