+++ src/main/java/com/dope/breaking/domain/user/Follow.java
@@ -1,8 +1,8 @@
-package com.dope.breaking.domain.UserRelationship;
+package com.dope.breaking.domain.user;
 
 
 
-import com.dope.breaking.domain.User;
+import com.dope.breaking.domain.user.User;
 
 import javax.persistence.*;
