+++ src/main/java/com/dope/breaking/domain/user/User.java
@@ -1,18 +1,20 @@
-package com.dope.breaking.domain;
+package com.dope.breaking.domain.user;
 
-import com.dope.breaking.domain.Financials.Purchase;
-import com.dope.breaking.domain.Financials.Statement;
-import com.dope.breaking.domain.UserRelationship.Block;
-import com.dope.breaking.domain.UserRelationship.Follow;
+import com.dope.breaking.domain.financial.Purchase;
+import com.dope.breaking.domain.financial.Statement;
+import com.dope.breaking.domain.post.Post;
+import lombok.Getter;
 
 import javax.persistence.*;
 import java.util.ArrayList;
 import java.util.List;
 
 @Entity
+@Getter
 public class User {
+
     @Id @GeneratedValue
-    @Column(name="USER_ID") //userId
+    @Column(name="USER_ID")
     private Long id;
 
     @OneToMany(mappedBy = "user")
@@ -26,8 +28,6 @@ public class User {
     @OneToMany(mappedBy = "following")
     private List<Follow> followerList = new ArrayList<Follow>();
 
-    //추후 차단 기능 구현//
-
     //유저가 차단한 리스트
     @OneToMany(mappedBy = "user")
     private List<Block> blockingList = new ArrayList<Block>();
@@ -36,30 +36,31 @@ public class User {
     @OneToMany(mappedBy = "blocking")
     private List<Block> blockerList = new ArrayList<Block>();
 
-    /////////////////////////////
-
-    //입출금 및 거래내역
-
+    //입출금 내역
     @OneToMany(mappedBy = "user")
     private List<Statement> statementList =  new ArrayList<Statement>();
-
+    //거래 내역
     @OneToMany(mappedBy = "user")
     private List<Purchase> purchaseList = new ArrayList<Purchase>();
 
-    /////////////////////////////
 
     @OneToMany(mappedBy = "user")
     private List<Bookmark> bookmarkList = new ArrayList<Bookmark>();
 
-
     private String firstname;
-    private String lastname;
-    private String nickname;
-    private String statusMsg;
-    private String email;
-    private String phoneNumber;
-    private int balance;
-    private String profileImgURL;
 
+    private String lastname;
+
+    private String nickname;
+
+    private String statusMsg;
+
+    private String email;
+
+    private String phoneNumber;
+
+    private int balance;
+
+    private String profileImgURL;
 
 }
